+++ CHANGELOG.md
@@ -7,12 +7,14 @@
 - Parameter `sourceName` changed to `source`
 - Parameter `entityName` changed to `entity`
 - Parameter `schemaName` changed to `schema`
-- Parameter `fileExpression` changed to `filter-expression`
+- Parameter `filterExpression` changed to `filter-expression`
 - Parameter `autoCreate` changed to `auto-create`
 - File Provider:
     - Changed `contentType` → `content` (Now supports multiple content type for the same storage, which may cause changes in behavior for existing configuration).
     - Changed `storageType` → `storage`.
     - Changed `fileSystem` storage → `fs`.
+    - Changed `azureBlob` storage → `azure-blob`.
+    - Changed Parameter `fsFolder` → `fs-folder`.
 - Plan sync function parameters have changed:
   - `source` → `from`
   - `destination` → `to`